+++ kikuchipy/signals/tests/test_ebsd.py
@@ -41,7 +41,6 @@ EMSOFT_FILE = os.path.join(DIR_PATH, "../../data/emsoft_ebsd/simulated_ebsd.h5")
 
 class TestEBSD:
     def test_init(self):
-
         # Signal shape
         array0 = np.zeros(shape=(10, 10, 10, 10))
         s0 = kp.signals.EBSD(array0)
@@ -261,7 +260,6 @@ class TestRemoveStaticBackgroundEBSD:
         """Test for expected error messages when passing an incorrect
         static background pattern to `remove_static_background().`
         """
-
         ebsd_node = kp.signals.util.metadata_nodes("ebsd")
         dummy_signal.metadata.set_item(
             ebsd_node + ".static_background", static_bg
@@ -282,11 +280,11 @@ class TestRemoveStaticBackgroundEBSD:
         dummy_signal2 = dummy_signal.deepcopy()
 
         dummy_signal.remove_static_background(
-            scale_bg=True, relative=False, static_bg=dummy_background,
+            scale_bg=True, relative=False, static_bg=dummy_background
         )
 
         dummy_signal2.remove_static_background(
-            scale_bg=False, relative=False, static_bg=dummy_background,
+            scale_bg=False, relative=False, static_bg=dummy_background
         )
 
         p1 = dummy_signal.inav[0, 0].data
@@ -302,7 +300,7 @@ class TestRemoveStaticBackgroundEBSD:
     ):
         with pytest.raises(ValueError, match="'scale_bg' must be False"):
             dummy_signal.remove_static_background(
-                relative=True, scale_bg=True, static_bg=dummy_background,
+                relative=True, scale_bg=True, static_bg=dummy_background
             )
 
     def test_non_square_patterns(self):
@@ -395,9 +393,8 @@ class TestRemoveDynamicBackgroundEBSD:
         to be made, these hard-coded answers will have to be
         recalculated for the tests to pass.
         """
-
         dummy_signal.remove_dynamic_background(
-            operation=operation, std=std, filter_domain="spatial",
+            operation=operation, std=std, filter_domain="spatial"
         )
         answer = answer.reshape((3,) * 4).astype(np.uint8)
         assert np.allclose(dummy_signal.data, answer)
@@ -465,7 +462,7 @@ class TestRemoveDynamicBackgroundEBSD:
 
         filter_domain = "frequency"
         dummy_signal.remove_dynamic_background(
-            operation=operation, std=std, filter_domain=filter_domain,
+            operation=operation, std=std, filter_domain=filter_domain
         )
 
         assert dummy_signal.data.dtype == dtype_out
@@ -529,7 +526,6 @@ class TestRescaleIntensityEBSD:
         to be made, these hard-coded answers will have to be
         recalculated for the tests to pass.
         """
-
         dummy_signal.rescale_intensity(relative=relative, dtype_out=dtype_out)
 
         assert dummy_signal.data.dtype == answer.dtype
@@ -576,14 +572,12 @@ class TestRescaleIntensityEBSD:
     def test_rescale_intensity_raises_in_range_percentiles(self, dummy_signal):
         with pytest.raises(ValueError, match="'percentiles' must be None"):
             dummy_signal.rescale_intensity(
-                in_range=(1, 254), percentiles=(1, 99),
+                in_range=(1, 254), percentiles=(1, 99)
             )
 
     def test_rescale_intensity_raises_in_range_relative(self, dummy_signal):
         with pytest.raises(ValueError, match="'in_range' must be None if "):
-            dummy_signal.rescale_intensity(
-                in_range=(1, 254), relative=True,
-            )
+            dummy_signal.rescale_intensity(in_range=(1, 254), relative=True)
 
 
 class TestAdaptiveHistogramEqualizationEBSD:
@@ -591,7 +585,6 @@ class TestAdaptiveHistogramEqualizationEBSD:
         """Test setup of equalization only. Tests of the result of the
         actual equalization are found elsewhere.
         """
-
         s = kp.load(KIKUCHIPY_FILE)
 
         # These window sizes should work without issue
@@ -675,18 +668,18 @@ class TestAverageNeighbourPatternsEBSD:
         ],
     )
     def test_average_neighbour_patterns(
-        self, dummy_signal, window, window_shape, lazy, answer, kwargs,
+        self, dummy_signal, window, window_shape, lazy, answer, kwargs
     ):
         if lazy:
             dummy_signal = dummy_signal.as_lazy()
 
         if kwargs is None:
             dummy_signal.average_neighbour_patterns(
-                window=window, window_shape=window_shape,
+                window=window, window_shape=window_shape
             )
         else:
             dummy_signal.average_neighbour_patterns(
-                window=window, window_shape=window_shape, **kwargs,
+                window=window, window_shape=window_shape, **kwargs
             )
 
         answer = answer.reshape((3, 3, 3, 3)).astype(np.uint8)
@@ -697,7 +690,7 @@ class TestAverageNeighbourPatternsEBSD:
         answer = dummy_signal.data.copy()
         with pytest.warns(UserWarning, match="A window of shape .* was "):
             dummy_signal.average_neighbour_patterns(
-                window="rectangular", window_shape=(1, 1),
+                window="rectangular", window_shape=(1, 1)
             )
         assert np.allclose(dummy_signal.data, answer)
         assert dummy_signal.data.dtype == answer.dtype
@@ -963,7 +956,7 @@ class TestGetDynamicBackgroundEBSD:
     def test_get_dynamic_background_spatial(self, dummy_signal):
         dtype_out = dummy_signal.data.dtype
         bg = dummy_signal.get_dynamic_background(
-            filter_domain="spatial", std=2, truncate=3,
+            filter_domain="spatial", std=2, truncate=3
         )
 
         assert bg.data.dtype == dtype_out
@@ -972,7 +965,7 @@ class TestGetDynamicBackgroundEBSD:
     def test_get_dynamic_background_frequency(self, dummy_signal):
         dtype_out = np.float32
         bg = dummy_signal.get_dynamic_background(
-            filter_domain="frequency", std=2, truncate=3, dtype_out=dtype_out,
+            filter_domain="frequency", std=2, truncate=3, dtype_out=dtype_out
         )
 
         assert bg.data.dtype == dtype_out
@@ -987,11 +980,9 @@ class TestGetDynamicBackgroundEBSD:
         lazy_signal = dummy_signal.as_lazy()
 
         bg = lazy_signal.get_dynamic_background()
-
         assert isinstance(bg, kp.signals.LazyEBSD)
 
         bg.compute()
-
         assert isinstance(bg, kp.signals.EBSD)
 
 
@@ -1030,7 +1021,6 @@ class TestGetImageQualityEBSD:
             dummy_signal = dummy_signal.as_lazy()
 
         iq = dummy_signal.get_image_quality(normalize=normalize)
-
         if lazy:
             iq = iq.compute()
 
@@ -1076,7 +1066,7 @@ class TestFFTFilterEBSD:
         w = kp.filters.Window(transfer_function, shape=shape, **kwargs)
 
         dummy_signal.fft_filter(
-            transfer_function=w, function_domain="frequency", shift=shift,
+            transfer_function=w, function_domain="frequency", shift=shift
         )
 
         assert isinstance(dummy_signal, kp.signals.EBSD)
@@ -1095,7 +1085,7 @@ class TestFFTFilterEBSD:
         w = np.array([[1, 0, -1], [2, 0, -2], [1, 0, -1]])
 
         dummy_signal.fft_filter(
-            transfer_function=w, function_domain="spatial", shift=False,
+            transfer_function=w, function_domain="spatial", shift=False
         )
         p2 = dummy_signal.inav[0, 0].data
         assert not np.allclose(p, p2, atol=1e-1)
@@ -1240,7 +1230,7 @@ class TestEBSDXmapProperty:
 
         # Should fail
         xmap_bad = get_single_phase_xmap(
-            nav_shape=nav_shape[::-1], step_sizes=step_sizes,
+            nav_shape=nav_shape[::-1], step_sizes=step_sizes
         )
         with pytest.raises(AttributeError, match="The crystal map shape"):
             s.xmap = xmap_bad
@@ -1275,19 +1265,26 @@ class TestEBSDDetectorProperty:
     def test_attribute_carry_over_from_deepcopy(self):
         s = kp.data.nickel_ebsd_small(lazy=True)
         s._detector = None
-        s2 = s.deepcopy()
+        _ = s.deepcopy()
 
 
-class TestPatternMatching:
-    def test_match_patterns(self, dummy_signal):
+class TestDictionaryIndexing:
+    def test_dictionary_indexing(self, dummy_signal):
         """Scores are all 1.0 for a dictionary containing all patterns
         from dummy_signal().
         """
         s_dict = kp.signals.EBSD(dummy_signal.data.reshape(-1, 3, 3))
         s_dict._xmap = CrystalMap(Rotation(np.zeros((9, 4))))
-        xmap = dummy_signal.match_patterns(s_dict)
+        xmap = dummy_signal.dictionary_indexing(s_dict)
 
         assert np.allclose(xmap.scores[:, 0], 1)
+
+    def test_match_patterns_warns(self, dummy_signal):
+        # TODO: Remove test after v0.4 is released
+        s_dict = kp.signals.EBSD(dummy_signal.data.reshape(-1, 3, 3))
+        s_dict._xmap = CrystalMap(Rotation(np.zeros((9, 4))))
+        with pytest.warns(np.VisibleDeprecationWarning, match="Function "):
+            _ = dummy_signal.match_patterns(s_dict)
 
 
 class TestAverageNeighbourDotProductMap:
@@ -1540,10 +1537,7 @@ class TestNeighbourDotProductMatrices:
         dp_matrices = s.get_neighbour_dot_product_matrices(window=window)
 
         assert np.allclose(
-            dp_matrices[1, 1],
-            desired_dp_matrices_11,
-            atol=1e-5,
-            equal_nan=True,
+            dp_matrices[1, 1], desired_dp_matrices_11, atol=1e-5, equal_nan=True
         )
 
     @pytest.mark.parametrize("dtype_out", [np.float16, np.float32, np.float64])
@@ -1579,7 +1573,7 @@ class TestNeighbourDotProductMatrices:
         dp_matrices = s.get_neighbour_dot_product_matrices(zero_mean=zero_mean)
 
         assert np.allclose(
-            dp_matrices[1, 1], desired_dp_matrices11, atol=1e-5, equal_nan=True,
+            dp_matrices[1, 1], desired_dp_matrices11, atol=1e-5, equal_nan=True
         )
 
     @pytest.mark.parametrize(
@@ -1608,7 +1602,7 @@ class TestNeighbourDotProductMatrices:
         dp_matrices = s.get_neighbour_dot_product_matrices(normalize=normalize)
 
         assert np.allclose(
-            dp_matrices[1, 1], desired_dp_matrices11, atol=1e-5, equal_nan=True,
+            dp_matrices[1, 1], desired_dp_matrices11, atol=1e-5, equal_nan=True
         )
 
     def test_dp_matrices_large(self):