+++ doc/changelog.rst
@@ -40,6 +40,12 @@ Changed
   colorbar only if desired.
   (`#375 <https://github.com/pyxem/kikuchipy/pull/375>`_)
 
+Deprecated
+----------
+- The EBSD.match_patterns() method changes name to EBSD.dictionary_indexing(),
+  and will be removed in v0.5.
+  (`#? <https://github.com/pyxem/kikuchipy/pull/?>`_)
+
 Fixed
 -----
 - Deep copying EBSD and EBSDMasterPattern signals carry over, respectively,